Split up gamepad.rs into smaller chunks
