+++ crates/bevy_input/src/gamepad/axis.rs
@@ -1,0 +1,5 @@
+use crate::gamepad::{Gamepad, GamepadAxisType};
+
+#[derive(Debug, Copy, Clone, PartialEq, Eq, Hash)]
+#[cfg_attr(feature = "serialize", derive(serde::Serialize, serde::Deserialize))]
+pub struct GamepadAxis(pub Gamepad, pub GamepadAxisType);
