+++ crates/bevy_input/src/gamepad/button.rs
@@ -1,0 +1,5 @@
+use crate::gamepad::{Gamepad, GamepadButtonType};
+
+#[derive(Debug, Copy, Clone, PartialEq, Eq, Hash)]
+#[cfg_attr(feature = "serialize", derive(serde::Serialize, serde::Deserialize))]
+pub struct GamepadButton(pub Gamepad, pub GamepadButtonType);
