+++ crates/bevy_input/src/gamepad/event.rs
@@ -1,0 +1,9 @@
+use crate::gamepad::{Gamepad, GamepadEventType};
+
+#[derive(Debug, Clone, PartialEq)]
+#[cfg_attr(feature = "serialize", derive(serde::Serialize, serde::Deserialize))]
+pub struct GamepadEvent(pub Gamepad, pub GamepadEventType);
+
+#[derive(Debug, Clone, PartialEq)]
+#[cfg_attr(feature = "serialize", derive(serde::Serialize, serde::Deserialize))]
+pub struct GamepadEventRaw(pub Gamepad, pub GamepadEventType);
