+++ crates/bevy_input/src/gamepad/gamepads.rs
@@ -1,0 +1,215 @@
+use bevy_utils::HashSet;
+
+#[derive(Debug, Copy, Clone, PartialEq, Eq, Hash)]
+#[cfg_attr(feature = "serialize", derive(serde::Serialize, serde::Deserialize))]
+pub struct Gamepad(pub usize);
+
+/// Container of unique connected [`Gamepad`]s
+///
+/// [`Gamepad`]s are registered and deregistered in [`gamepad_connection_system`]
+#[derive(Default)]
+pub struct Gamepads {
+    gamepads: HashSet<Gamepad>,
+}
+
+impl Gamepads {
+    /// Returns true if the [Gamepads] contains a [Gamepad].
+    pub fn contains(&self, gamepad: &Gamepad) -> bool {
+        self.gamepads.contains(gamepad)
+    }
+
+    /// Iterates over registered [Gamepad]s
+    pub fn iter(&self) -> impl Iterator<Item = &Gamepad> + '_ {
+        self.gamepads.iter()
+    }
+
+    /// Registers [Gamepad].
+    pub(crate) fn register(&mut self, gamepad: Gamepad) {
+        self.gamepads.insert(gamepad);
+    }
+
+    /// Deregisters [Gamepad.
+    pub(crate) fn deregister(&mut self, gamepad: &Gamepad) {
+        self.gamepads.remove(gamepad);
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use crate::gamepad::{AxisSettings, ButtonAxisSettings, ButtonSettings};
+
+    fn test_button_axis_settings_filter(
+        settings: ButtonAxisSettings,
+        new_value: f32,
+        old_value: Option<f32>,
+        expected: Option<f32>,
+    ) {
+        let actual = settings.filter(new_value, old_value);
+        assert_eq!(
+            expected, actual,
+            "Testing filtering for {:?} with new_value = {:?}, old_value = {:?}",
+            settings, new_value, old_value
+        );
+    }
+
+    #[test]
+    fn test_button_axis_settings_default_filter() {
+        let cases = [
+            (1.0, None, Some(1.0)),
+            (0.99, None, Some(1.0)),
+            (0.96, None, Some(1.0)),
+            (0.95, None, Some(1.0)),
+            (0.9499, None, Some(0.9499)),
+            (0.84, None, Some(0.84)),
+            (0.43, None, Some(0.43)),
+            (0.05001, None, Some(0.05001)),
+            (0.05, None, Some(0.0)),
+            (0.04, None, Some(0.0)),
+            (0.01, None, Some(0.0)),
+            (0.0, None, Some(0.0)),
+        ];
+
+        for (new_value, old_value, expected) in cases {
+            let settings = ButtonAxisSettings::default();
+            test_button_axis_settings_filter(settings, new_value, old_value, expected);
+        }
+    }
+
+    #[test]
+    fn test_button_axis_settings_default_filter_with_old_value() {
+        let cases = [
+            (0.43, Some(0.44001), Some(0.43)),
+            (0.43, Some(0.44), None),
+            (0.43, Some(0.43), None),
+            (0.43, Some(0.41999), Some(0.43)),
+            (0.43, Some(0.17), Some(0.43)),
+            (0.43, Some(0.84), Some(0.43)),
+            (0.05, Some(0.055), Some(0.0)),
+            (0.95, Some(0.945), Some(1.0)),
+        ];
+
+        for (new_value, old_value, expected) in cases {
+            let settings = ButtonAxisSettings::default();
+            test_button_axis_settings_filter(settings, new_value, old_value, expected);
+        }
+    }
+
+    fn test_axis_settings_filter(
+        settings: AxisSettings,
+        new_value: f32,
+        old_value: Option<f32>,
+        expected: Option<f32>,
+    ) {
+        let actual = settings.filter(new_value, old_value);
+        assert_eq!(
+            expected, actual,
+            "Testing filtering for {:?} with new_value = {:?}, old_value = {:?}",
+            settings, new_value, old_value
+        );
+    }
+
+    #[test]
+    fn test_axis_settings_default_filter() {
+        let cases = [
+            (1.0, Some(1.0)),
+            (0.99, Some(1.0)),
+            (0.96, Some(1.0)),
+            (0.95, Some(1.0)),
+            (0.9499, Some(0.9499)),
+            (0.84, Some(0.84)),
+            (0.43, Some(0.43)),
+            (0.05001, Some(0.05001)),
+            (0.05, Some(0.0)),
+            (0.04, Some(0.0)),
+            (0.01, Some(0.0)),
+            (0.0, Some(0.0)),
+            (-1.0, Some(-1.0)),
+            (-0.99, Some(-1.0)),
+            (-0.96, Some(-1.0)),
+            (-0.95, Some(-1.0)),
+            (-0.9499, Some(-0.9499)),
+            (-0.84, Some(-0.84)),
+            (-0.43, Some(-0.43)),
+            (-0.05001, Some(-0.05001)),
+            (-0.05, Some(0.0)),
+            (-0.04, Some(0.0)),
+            (-0.01, Some(0.0)),
+        ];
+
+        for (new_value, expected) in cases {
+            let settings = AxisSettings::default();
+            test_axis_settings_filter(settings, new_value, None, expected);
+        }
+    }
+
+    #[test]
+    fn test_axis_settings_default_filter_with_old_values() {
+        let cases = [
+            (0.43, Some(0.44001), Some(0.43)),
+            (0.43, Some(0.44), None),
+            (0.43, Some(0.43), None),
+            (0.43, Some(0.41999), Some(0.43)),
+            (0.43, Some(0.17), Some(0.43)),
+            (0.43, Some(0.84), Some(0.43)),
+            (0.05, Some(0.055), Some(0.0)),
+            (0.95, Some(0.945), Some(1.0)),
+            (-0.43, Some(-0.44001), Some(-0.43)),
+            (-0.43, Some(-0.44), None),
+            (-0.43, Some(-0.43), None),
+            (-0.43, Some(-0.41999), Some(-0.43)),
+            (-0.43, Some(-0.17), Some(-0.43)),
+            (-0.43, Some(-0.84), Some(-0.43)),
+            (-0.05, Some(-0.055), Some(0.0)),
+            (-0.95, Some(-0.945), Some(-1.0)),
+        ];
+
+        for (new_value, old_value, expected) in cases {
+            let settings = AxisSettings::default();
+            test_axis_settings_filter(settings, new_value, old_value, expected);
+        }
+    }
+
+    #[test]
+    fn test_button_settings_default_is_pressed() {
+        let cases = [
+            (1.0, true),
+            (0.95, true),
+            (0.9, true),
+            (0.8, true),
+            (0.75, true),
+            (0.7, false),
+            (0.65, false),
+            (0.5, false),
+            (0.0, false),
+        ];
+
+        for (value, expected) in cases {
+            let settings = ButtonSettings::default();
+            let actual = settings.is_pressed(value);
+
+            assert_eq!(expected, actual, "Testing is pressed for value: {}", value);
+        }
+    }
+
+    #[test]
+    fn test_button_settings_default_is_released() {
+        let cases = [
+            (1.0, false),
+            (0.95, false),
+            (0.9, false),
+            (0.8, false),
+            (0.75, false),
+            (0.7, false),
+            (0.65, true),
+            (0.5, true),
+            (0.0, true),
+        ];
+
+        for (value, expected) in cases {
+            let settings = ButtonSettings::default();
+            let actual = settings.is_released(value);
+
+            assert_eq!(expected, actual, "Testing is released for value: {}", value);
+        }
+    }
+}
