+++ crates/bevy_input/src/gamepad/settings.rs
@@ -1,0 +1,146 @@
+use crate::gamepad::{GamepadAxis, GamepadButton};
+use bevy_utils::HashMap;
+
+#[derive(Default, Debug)]
+pub struct GamepadSettings {
+    pub default_button_settings: ButtonSettings,
+    pub default_axis_settings: AxisSettings,
+    pub default_button_axis_settings: ButtonAxisSettings,
+    pub button_settings: HashMap<GamepadButton, ButtonSettings>,
+    pub axis_settings: HashMap<GamepadAxis, AxisSettings>,
+    pub button_axis_settings: HashMap<GamepadButton, ButtonAxisSettings>,
+}
+
+impl GamepadSettings {
+    pub fn get_button_settings(&self, button: GamepadButton) -> &ButtonSettings {
+        self.button_settings
+            .get(&button)
+            .unwrap_or(&self.default_button_settings)
+    }
+
+    pub fn get_axis_settings(&self, axis: GamepadAxis) -> &AxisSettings {
+        self.axis_settings
+            .get(&axis)
+            .unwrap_or(&self.default_axis_settings)
+    }
+
+    pub fn get_button_axis_settings(&self, button: GamepadButton) -> &ButtonAxisSettings {
+        self.button_axis_settings
+            .get(&button)
+            .unwrap_or(&self.default_button_axis_settings)
+    }
+}
+
+#[derive(Debug, Clone)]
+pub struct ButtonSettings {
+    pub press: f32,
+    pub release: f32,
+}
+
+impl Default for ButtonSettings {
+    fn default() -> Self {
+        ButtonSettings {
+            press: 0.75,
+            release: 0.65,
+        }
+    }
+}
+
+impl ButtonSettings {
+    pub(crate) fn is_pressed(&self, value: f32) -> bool {
+        value >= self.press
+    }
+
+    pub(crate) fn is_released(&self, value: f32) -> bool {
+        value <= self.release
+    }
+}
+
+/// Defines the sensitivity range and threshold for an axis.
+///
+/// Values that are lower than `negative_high` will be rounded to -1.0.
+/// Values that are higher than `positive_high` will be rounded to 1.0.
+/// Values that are in-between `negative_low` and `positive_low` will be rounded to 0.0.
+/// Otherwise, values will not be rounded.
+///
+/// The valid range is from -1.0 to 1.0, inclusive.
+#[derive(Debug, Clone)]
+pub struct AxisSettings {
+    pub positive_high: f32,
+    pub positive_low: f32,
+    pub negative_high: f32,
+    pub negative_low: f32,
+    ///`threshold` defines the minimum difference between old and new values to apply the changes.
+    pub threshold: f32,
+}
+
+impl Default for AxisSettings {
+    fn default() -> Self {
+        AxisSettings {
+            positive_high: 0.95,
+            positive_low: 0.05,
+            negative_high: -0.95,
+            negative_low: -0.05,
+            threshold: 0.01,
+        }
+    }
+}
+
+impl AxisSettings {
+    pub(crate) fn filter(&self, new_value: f32, old_value: Option<f32>) -> Option<f32> {
+        let new_value = if new_value <= self.positive_low && new_value >= self.negative_low {
+            0.0
+        } else if new_value >= self.positive_high {
+            1.0
+        } else if new_value <= self.negative_high {
+            -1.0
+        } else {
+            new_value
+        };
+
+        if let Some(old_value) = old_value {
+            if (new_value - old_value).abs() <= self.threshold {
+                return None;
+            }
+        }
+
+        Some(new_value)
+    }
+}
+
+#[derive(Debug, Clone)]
+pub struct ButtonAxisSettings {
+    pub high: f32,
+    pub low: f32,
+    pub threshold: f32,
+}
+
+impl Default for ButtonAxisSettings {
+    fn default() -> Self {
+        ButtonAxisSettings {
+            high: 0.95,
+            low: 0.05,
+            threshold: 0.01,
+        }
+    }
+}
+
+impl ButtonAxisSettings {
+    pub(crate) fn filter(&self, new_value: f32, old_value: Option<f32>) -> Option<f32> {
+        let new_value = if new_value <= self.low {
+            0.0
+        } else if new_value >= self.high {
+            1.0
+        } else {
+            new_value
+        };
+
+        if let Some(old_value) = old_value {
+            if (new_value - old_value).abs() <= self.threshold {
+                return None;
+            }
+        }
+
+        Some(new_value)
+    }
+}
